# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Restaurant API",
        "description": "Dokumentasi API autentikasi untuk aplikasi restoran.",
        "version": "1.0"
    },
    "paths": {
...
        "/api/tables": {
            "get": {
                "tags": [
                    "Tables"
                ],
                "summary": "Dapatkan semua meja",
                "operationId": "ff9d243a2aac0c783b0b20f5bd9d5913",
                "responses": {
                    "200": {
                        "description": "List meja berhasil diambil"
                    }
                }
            },
            "post": {
                "tags": [
                    "Tables"
                ],
                "summary": "Buat meja baru dengan QR code",
                "operationId": "ee3cdb4f76abd42e86f037b17e4d0206",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "table_number",
                                    "capacity"
                                ],
                                "properties": {
                                    "table_number": {
                                        "type": "string",
                                        "example": "1"
                                    },
                                    "capacity": {
                                        "type": "integer",
                                        "example": 4
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "Indoor",
                                            "Outdoor",
                                            "VIP"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Meja berhasil dibuat dengan QR code"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/tables/{id}": {
            "get": {
                "tags": [
                    "Tables"
                ],
                "summary": "Dapatkan detail meja",
                "operationId": "465490c7477f4f985439a763800c841c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail meja"
                    },
                    "404": {
                        "description": "Meja tidak ditemukan"
                    }
                }
            },
            "put": {
                "tags": [
                    "Tables"
                ],
                "summary": "Update meja",
                "operationId": "1b0db842a021f0e5558f861a6dfc961a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "table_number": {
                                        "type": "string"
                                    },
                                    "capacity": {
                                        "type": "integer"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "available",
                                            "occupied",
                                            "reserved"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Meja berhasil diupdate"
                    },
                    "404": {
                        "description": "Meja tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Tables"
                ],
                "summary": "Hapus meja",
                "operationId": "fe0835f1e615253c8ba068fad5293f97",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Meja berhasil dihapus"
                    },
                    "404": {
                        "description": "Meja tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}